refactor: split features/tenant/security/policy_engine.go into 3 files (Issue #455)
policy_engine.go (1,233 lines) mixed policy evaluation core, zero-trust
overlay coordination (4 modes), audit logging, and conflict resolution in
one file. The zero-trust coordination alone had 300+ lines of mode-specific
logic that was difficult to navigate independently of the core evaluation
logic.

Extracted two new files to separate the responsibility clusters:

- policy_audit.go (~95 lines): LogSecurityPolicyChange, LogPolicyEvaluation,
  LogZeroTrustOverlayEvaluation — fire-and-forget audit methods with no
  coupling to coordination logic.

- policy_coordination.go (~485 lines): TenantZeroTrustMode constants,
  TenantPolicyCoordination and related coordination types, ZeroTrustOverlayResult,
  NewDefaultTenantPolicyCoordination, evaluateZeroTrustOverlay,
  coordinatePolicyResults, convertToZeroTrustRequest, calculatePolicyAlignment,
  determineOverlayAction, determineIntegratedAction, buildOverlayReason,
  applyOverlayMode (enforced/governing/integrated/monitoring),
  applyConflictResolution.

- policy_engine.go (trimmed to ~678 lines): Core types (TenantSecurityPolicy,
  rule definitions, evaluation request/result types), compiled rule
  implementations (BaseRule, CompiledDataSecurityRule, CompiledAccessControlRule),
  and the engine methods (CreateSecurityPolicy, EvaluateSecurityPolicy,
  evaluateTenantSecurityPolicy, compileSecurityRules).

Changes:
- Extract audit methods to policy_audit.go (no logic changes)
- Extract coordination types and methods to policy_coordination.go
- Trim policy_engine.go from 1,233 to 678 lines (~45% reduction)
- All new files are under 500 lines
- Zero public API surface changes

Test results (features/tenant/security): all tests pass.
enhanced_multi_tenant_security_test.go passes unchanged.

Fixes #455
